Implement category relation to theme form

+++ src/views/Theme/Theme.vue
@@ -10,9 +10,9 @@
             <div class="container-fluid d-flex align-items-center">
                 <div class="row">
                     <div class="col-lg-7 col-md-10">
-                        <h1 class="display-2 text-white">{{ category ? category.name : "" }}</h1>
+                        <h1 class="display-2 text-white">{{ themeModel.name ? themeModel.name : "New Theme" }}</h1>
 
-                        <p class="text-white mt-0 mb-5">{{ category ? category.description : "" }}</p>
+                        <p class="text-white mt-0 mb-5">{{ themeModel.description ? themeModel.description : "" }}</p>
                     </div>
                 </div>
             </div>
@@ -72,45 +72,79 @@
                                 </div>
                             </div>
 
+                            <div class="row">
+                                <div class="col-lg-6">
+                                    <label class="font-weight-bold"> Authors </label>
+
+                                    <vue-tags-input
+                                        v-model="authorTagModel"
+                                        :tags="authorSelectedTags"
+                                        :allow-edit-tags="false"
+                                        :autocomplete-items="filteredAuthorTagItems"
+                                        :avoid-adding-duplicates="true"
+                                        @tags-changed="authorTagsChanged"
+                                        @before-adding-tag="beforeAddingAuthorTag"
+                                    >
+                                        <template v-slot:autocomplete-item="props">
+                                            <div class="row mx-2 py-1">
+                                                <div class="col-lg-1">
+                                                    <img
+                                                        class="rounded"
+                                                        :src="props.item.avatar ? props.item.avatar : 'img/theme/default-placeholder.png'"
+                                                        alt="Author Avatar"
+                                                        height="30"
+                                                        width="30"
+                                                    />
+                                                </div>
+
+                                                <div class="col-lg-11">
+                                                    <label @click="props.performAdd(props.item)" class="w-100 px-2 mt-1">
+                                                        {{ props.item.name }}
+                                                    </label>
+                                                </div>
+                                            </div>
+                                        </template>
+                                    </vue-tags-input>
+                                </div>
+
+                                <div class="col-lg-6">
+                                    <label class="font-weight-bold"> Categories </label>
+
+                                    <vue-tags-input
+                                        v-model="categoryTagModel"
+                                        :tags="categorySelectedTags"
+                                        :allow-edit-tags="false"
+                                        :autocomplete-items="filteredCategoryTagItems"
+                                        :avoid-adding-duplicates="true"
+                                        @tags-changed="categoryTagsChanged"
+                                        @before-adding-tag="beforeAddingCategoryTag"
+                                    >
+                                        <template v-slot:autocomplete-item="props">
+                                            <div class="row mx-2 py-1">
+                                                <div class="col-lg-1">
+                                                    <img
+                                                        class="rounded"
+                                                        :src="props.item.image ? props.item.image : 'img/theme/default-placeholder.png'"
+                                                        alt="Category Image"
+                                                        height="30"
+                                                        width="30"
+                                                    />
+                                                </div>
+
+                                                <div class="col-lg-11">
+                                                    <label @click="props.performAdd(props.item)" class="w-100 px-2 mt-1">
+                                                        {{ props.item.name }}
+                                                    </label>
+                                                </div>
+                                            </div>
+                                        </template>
+                                    </vue-tags-input>
+                                </div>
+                            </div>
+
                             <div class="row mt-3">
                                 <div class="col">
                                     <div class="row">
-                                        <div class="col-lg-12">
-                                            <label class="font-weight-bold"> Authors </label>
-
-                                            <vue-tags-input
-                                                v-model="inputTagModel"
-                                                :tags="inputSelectedTags"
-                                                :allow-edit-tags="false"
-                                                :autocomplete-items="filteredItems"
-                                                :avoid-adding-duplicates="true"
-                                                @tags-changed="tagsChanged"
-                                                @before-adding-tag="beforeAddingTag"
-                                            >
-                                                <template v-slot:autocomplete-item="props">
-                                                    <div class="row mx-2 py-1">
-                                                        <div class="col-lg-1">
-                                                            <img
-                                                                class="rounded"
-                                                                :src="props.item.avatar ? props.item.avatar : 'img/theme/default-placeholder.png'"
-                                                                alt="Author Avatar"
-                                                                height="30"
-                                                                width="30"
-                                                            />
-                                                        </div>
-
-                                                        <div class="col-lg-11">
-                                                            <label @click="props.performAdd(props.item)" class="w-100 px-2 mt-1">
-                                                                {{ props.item.name }}
-                                                            </label>
-                                                        </div>
-                                                    </div>
-                                                </template>
-                                            </vue-tags-input>
-                                        </div>
-                                    </div>
-
-                                    <div class="row mt-4">
                                         <div class="col-lg-12">
                                             <label class="font-weight-bold"> Image </label>
 
@@ -151,6 +185,7 @@ import ThemeService from "@/services/ThemeService";
 import AuthorService from "@/services/AuthorService";
 import NotificationService from "@/services/NotificationService";
 import FilepondService from "@/services/FilepondService";
+import CategoryService from "@/services/CategoryService";
 
 export default {
     name: "theme-page",
@@ -161,6 +196,8 @@ export default {
         return {
             pond: null,
 
+            themeId: null,
+
             themeModel: {
                 name: "",
                 description: "",
@@ -168,34 +205,61 @@ export default {
                 group: "",
                 status: true,
                 authors: [],
+                categories: [],
             },
 
-            inputTagModel: "",
-            inputSelectedTags: [],
-
+            authorTagModel: "",
+            authorSelectedTags: [],
             authorsList: [],
+
+            categoryTagModel: "",
+            categorySelectedTags: [],
+            categoryList: [],
         };
     },
 
     computed: {
-        filteredItems() {
-            return this.authorsList && this.authorsList.length
-                ? this.authorsList.filter((x) => {
-                      return x.name.toLowerCase().indexOf(this.inputTagModel.toLowerCase()) !== -1;
-                  })
-                : [];
+        filteredAuthorTagItems() {
+            if (this.authorsList && this.authorsList.length) {
+                const sugessions = this.authorsList.filter((x) => {
+                    return x.name.toLowerCase().indexOf(this.authorTagModel.toLowerCase()) !== -1;
+                });
+
+                return sugessions.length ? sugessions : this.authorsList.slice(0, 10);
+            }
+
+            return [];
+        },
+
+        filteredCategoryTagItems() {
+            if (this.categoryList && this.categoryList.length) {
+                const sugessions = this.categoryList.filter((x) => {
+                    return x.name.toLowerCase().indexOf(this.categoryTagModel.toLowerCase()) !== -1;
+                });
+
+                return sugessions.length ? sugessions : this.categoryList.slice(0, 10);
+            }
+
+            return [];
         },
     },
 
     mounted() {
-        this.initFilepond();
+        this.themeId = this.$route.params.id;
 
         this.loadAuthors();
+        this.loadCategories();
 
-        this.getTheme(true);
+        this.initFilepond();
+
+        if (this.themeId && this.themeId != "new") {
+            this.getTheme(this.themeId);
+        }
     },
 
     methods: {
+        // #region Author
+
         // load all authors for tags input
         loadAuthors() {
             AuthorService.getAll(0, true).then((response) => {
@@ -207,8 +271,13 @@ export default {
             });
         },
 
+        // tags changed event
+        authorTagsChanged(newTags) {
+            this.authorSelectedTags = newTags;
+        },
+
         // before adding author tag
-        beforeAddingTag(tagObject) {
+        beforeAddingAuthorTag(tagObject) {
             (this.authorsList && this.authorsList.length
                 ? this.authorsList.filter((x) => {
                       return x.name === tagObject.tag.name;
@@ -218,41 +287,80 @@ export default {
                 : null;
         },
 
-        // tags changed event
-        tagsChanged(newTags) {
-            this.inputSelectedTags = newTags;
+        // #endregion
+
+        // #region Category
+
+        // load all categories for tags input
+        loadCategories() {
+            CategoryService.getAll(0, true).then((response) => {
+                this.categoryList = response.data.result.map(function (item) {
+                    item.text = item.name;
+
+                    return item;
+                });
+            });
         },
+
+        // tags changed event
+        categoryTagsChanged(newTags) {
+            this.categorySelectedTags = newTags;
+        },
+
+        // before adding category tag
+        beforeAddingCategoryTag(tagObject) {
+            (this.categoryList && this.categoryList.length
+                ? this.categoryList.filter((x) => {
+                      return x.name === tagObject.tag.name;
+                  }).length
+                : 0) > 0
+                ? tagObject.addTag()
+                : null;
+        },
+
+        // #endregion
 
         // setup and create filepond instance
         initFilepond() {
-            const token = store.get("userAccessToken");
+            if (!this.themeId || this.themeId == "new") {
+                const token = store.get("userAccessToken");
 
-            FilepondService.setOptions({
-                server: {
-                    url: ThemeService.getImageUploadEndpoint(this.$route.params.id),
-                    process: {
-                        headers: {
-                            Authorization: `Bearer ${token}`,
+                FilepondService.setOptions({
+                    server: {
+                        url: ThemeService.getImageUploadEndpoint(this.themeId),
+                        process: {
+                            headers: {
+                                Authorization: `Bearer ${token}`,
+                            },
                         },
                     },
-                },
-            });
+                });
+            }
 
             this.pond = FilepondService.create();
 
             document.querySelector("#pond-container").appendChild(this.pond.element);
         },
 
-        getTheme() {
-            ThemeService.get(this.$route.params.id).then((response) => {
+        getTheme(id) {
+            ThemeService.get(id).then((response) => {
                 this.themeModel = response.data.result;
+
                 this.themeModel.authors = response.data.result.authors.map((item) => {
                     item.text = item.name;
 
                     return item;
                 });
 
-                this.inputSelectedTags = this.themeModel.authors;
+                this.authorSelectedTags = this.themeModel.authors;
+
+                this.themeModel.categories = response.data.result.categories.map((item) => {
+                    item.text = item.name;
+
+                    return item;
+                });
+
+                this.categorySelectedTags = this.themeModel.categories;
             });
         },
 
@@ -260,25 +368,65 @@ export default {
         submitThemeForm(e) {
             e.preventDefault();
 
-            const authors = JSON.parse(JSON.stringify(this.inputSelectedTags)).map(function (item) {
+            const authors = JSON.parse(JSON.stringify(this.authorSelectedTags)).map(function (item) {
+                return item.id;
+            });
+
+            const categories = JSON.parse(JSON.stringify(this.categorySelectedTags)).map(function (item) {
                 return item.id;
             });
 
             this.themeModel.authors = authors;
+            this.themeModel.categories = categories;
 
-            ThemeService.update(this.$route.params.id, this.themeModel).then(() => {
-                this.pond.processFile().then(() => {
-                    this.themeModel.cover = null;
+            if (this.themeId && this.themeId != "new") {
+                ThemeService.update(this.themeId, this.themeModel).then(() => {
+                    this.pond.processFile().then(() => {
+                        this.themeModel.cover = null;
 
-                    this.pond.removeFile();
+                        this.pond.removeFile();
+
+                        this.getTheme();
+                    });
 
                     this.getTheme();
+
+                    NotificationService.success("Theme updated successfully");
                 });
+            } else {
+                ThemeService.create(this.themeModel)
+                    .then((response) => {
+                        const themeId = response.data.result.id,
+                            token = store.get("userAccessToken");
 
-                this.getTheme();
+                        FilepondService.setOptions({
+                            server: {
+                                url: ThemeService.getImageUploadEndpoint(themeId),
+                                process: {
+                                    headers: {
+                                        Authorization: `Bearer ${token}`,
+                                    },
+                                },
+                            },
+                        });
 
-                NotificationService.success("Theme updated successfully");
-            });
+                        this.themeId = themeId;
+
+                        this.pond.processFile();
+
+                        NotificationService.success("Theme created successfully");
+                    })
+                    .then(() => {
+                        this.$router.push({
+                            name: "theme",
+                            params: {
+                                id: this.themeId,
+                            },
+                        });
+
+                        this.getTheme(this.themeId);
+                    });
+            }
         },
     },
 };
